+++ src/program/buildSetupCommand.ts
@@ -12,11 +12,21 @@ export const buildSetupCommand = (program: Command, setup: SetupCommand) => {
     .command('setup')
     .description('Setup the project for Expo ICP integration')
     .option(
+      '--expo-crypto-universal',
+      'Install Expo Crypto Universal packages',
+      false,
+    )
+    .option(
+      '--expo-storage-universal',
+      'Install Expo Storage Universal packages',
+      false,
+    )
+    .option('--expo-icp-helpers', 'Install Expo ICP helper packages', false)
+    .option(
       '--ii-integration-helpers',
       'Install II integration helper packages',
       false,
     )
-    .option('--expo-icp-helpers', 'Install Expo ICP helper packages', false)
     .option(
       '--expo-icp-app-connect',
       'Install Expo ICP app connect packages',